+++ packages/docs/src/stories/Box.stories.tsx
@@ -4,7 +4,7 @@ import { Box, Text } from '@gibbon-design/react'
 const meta: Meta<typeof Box> = {
   title: 'Surfaces/Box',
   component: Box,
-  tags: ['autodocs'],// criam um arquivo para mostrar as docs
+  tags: ['autodocs'],
   args: {
     children: <Text> Testando o elemento Box </Text>
   },
@@ -21,3 +21,9 @@ export default meta
 type Story = StoryObj<typeof Box>
 
 export const Primary: Story = {}
+
+export const CustomTag: Story = {
+  args: {
+    children: <Text as="strong">Strong text</Text>
+  }
+}