+++ packages/docs/src/stories/TextArea.stories.tsx
@@ -9,7 +9,7 @@ const meta: Meta<typeof TextArea> = {
   decorators: [
     (Story) => {
       return (
-        <Box as="label" css={{ display: 'flex', flexDirection: 'column', gap: '$2'}}>
+        <Box as="label" style={{ display: 'flex', flexDirection: 'column', gap: '$2'}}>
           <Text size="sm" >Observations</Text>
           {Story()}
         </Box>