+++ packages/react/src/components/Avatar/styles.ts
@@ -1,28 +1,37 @@
-import * as Avatar from '@radix-ui/react-avatar'
-import { styled } from '../../styles'
-export const AvatarContainer = styled(Avatar.Root, {
-  borderRadius: '$full',
-  display: 'inline-block',
-  width: '$12',
-  height: '$12',
-  overflow: 'hidden',
-})
-export const AvatarImage = styled(Avatar.Image, {
-  width: '100%',
-  height: '100%',
-  objectFit: 'cover',
-  borderRadius: 'inherit',
-})
-export const AvatarFallback = styled(Avatar.Fallback, {
-  width: '100%',
-  height: '100%',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  backgroundColor: '$gray600',
-  color: '$gray800',
-  svg: {
-    width: '$6',
-    height: '$6',
-  },
-})
+import styled from 'styled-components';
+import * as Avatar from '@radix-ui/react-avatar';
+import { theme } from '../../styles/themes';
+
+
+// Container do Avatar
+export const AvatarContainer = styled(Avatar.Root)`
+  border-radius: ${ theme.radii.full };
+  display: inline-block;
+  width: ${ theme.space[12] };
+  height: ${ theme.space[12] };
+  overflow: hidden;
+`;
+
+// Imagem do Avatar
+export const AvatarImage = styled(Avatar.Image)`
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: inherit;
+`;
+
+// Fallback do Avatar
+export const AvatarFallback = styled(Avatar.Fallback)`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: ${ theme.colors.gray600 };
+  color: ${ theme.colors.gray800 };
+
+  svg {
+    width: ${ theme.space[6] };
+    height: ${ theme.space[6] };
+  }
+`;