+++ packages/react/src/components/Box.tsx
@@ -1,8 +1,9 @@
-import { styled } from '../styles'
+import styled from "styled-components"
+import { theme } from "../styles/themes"
 
-export const Box = styled('div', {
-  padding: '$4',
-  borderRadius: '$md',
-  border: '1px solid $gray600',
-  backgroundColor: '$gray800',
-}) 
+export const Box = styled.div`
+  padding: ${theme.space[4]};
+  border-radius: ${theme.radii.md};
+  border: 1px solid ${theme.colors.gray600};
+  background-color: ${theme.colors.gray800};
+`