+++ packages/react/src/components/Button.tsx
@@ -1,78 +1,92 @@
-import { styled } from '../styles'
+import styled, { css } from "styled-components"
+import { theme } from "../styles/themes"
 
-export const Button = styled('button', {
-  all: 'unset',
-  borderRadius: '$sm',
-  fontSize: '$sm',
-  fontWeight: '$medium',
-  fontFamily: '$default',
-  textAlign: 'center',
-  minWidth: 120,
-  boxSizing: 'border-box',
-  padding: '0 $4',
+const stylesVariants = {
+  primary: css`
+    color: ${theme.colors.white};
+    background: ${theme.colors.green500};
+    
+    &:not(:disabled):hover {
+      background: ${theme.colors.green300};
+    };
 
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  gap: '$2',
+    &:disabled {
+      background-color: ${theme.colors.gray200};
+    }
+  `,
 
-  cursor: 'pointer',
+  secondary: css`
+    color: ${theme.colors.green300};
+    border: 2px solid ${theme.colors.green500};
+    
+    &:not(:disabled):hover {
+      background: ${theme.colors.green500};
+      color: ${theme.colors.white};
+    };
+
+    &:disabled {
+      color: ${theme.colors.gray200};
+      border-color: ${theme.colors.gray200};
+    };
+  `,
+  tertiary: css`
+    color: ${theme.colors.gray100};
+    
+    &:not(:disabled):hover {
+      color: ${theme.colors.white};
+    };
+
+    &:disabled {
+      color: ${theme.colors.gray600};
+    };
+  `,
+}
+
+const sizeVariants = {
+  sm: css`
+    height: 38px;
+  `,
+
+  md: css`
+    height: 46px;
+  `
+}  
+
+export const Button = styled.button<{
+  variant: keyof typeof stylesVariants;
+  size: keyof typeof sizeVariants;
+}>`
+  all: unset;
+  border-radius: ${theme.radii.sm};
+  font-size: ${theme.fontSizes.sm};
+  font-weight: ${theme.fontWeights.medium};
+  font-family: ${theme.fonts.default};
+  text-align: center;
+  min-width: 120px;
+  box-sizing: border-box;
+  padding: 0 ${theme.space[4]};
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: ${theme.space[2]};
+
+  cursor: pointer;
 
   svg: {
-    width: '$4',
-    height: '$4',
-  },
+    width: ${theme.space[4]};
+    height: ${theme.space[4]};
+  }
 
-  '&:disabled': {
-    cursor: 'not-allowed',
-  },
-  variants: {
-    variant: {
-      primary: {
-        color: '$white',
-        background: '$ignite500',
-        '&:not(:disabled):hover': {
-          background: '$ignite300',
-        },
-        '&:disabled': {
-          backgroundColor: '$gray200',
-        },
-      },
-      secondary: {
-        color: '$ignite300',
-        border: '2px solid $ignite500',
-        '&:not(:disabled):hover': {
-          background: '$ignite500',
-          color: '$white',
-        },
-        '&:disabled': {
-          color: '$gray200',
-          borderColor: '$gray200',
-        },
-      },
-      tertiary: {
-        color: '$gray100',
-        '&:not(:disabled):hover': {
-          color: '$white',
-        },
-        '&:disabled': {
-          color: '$gray600',
-        },
-      },
-    },
-    size: {
-      sm: {
-        height: 38,
-      },
+  &:disabled {
+    cursor: not-allowed;
+  }
 
-      md: {
-        height: 46,
-      },
-    },
-  },
+  ${({variant}) => stylesVariants[variant]}
+  ${({size}) => sizeVariants[size]}
+`
 
-  defaultVariants: {
-    variant: 'primary',
-    size: 'md',
-  },
-})
+Button.defaultProps = {
+  variant: 'primary',
+  size: 'md'
+}