+++ packages/react/src/components/CheckBox/index.tsx
@@ -1,16 +1,16 @@
-// import { Check } from '@phosphor-icons/react'
-// import { CheckboxContainer, CheckboxIndicator } from './styles'
-// import { ComponentProps } from 'react'
+import { Check } from '@phosphor-icons/react'
+import { CheckboxContainer, CheckboxIndicator } from './styles'
+import { ComponentProps } from 'react'
 
-// export interface CheckboxProps
-//   extends ComponentProps<typeof CheckboxContainer> {}
+export interface CheckboxProps
+  extends ComponentProps<typeof CheckboxContainer> {}
 
-// export function Checkbox(props: CheckboxProps) {
-//   return (
-//     <CheckboxContainer {...props}>
-//       <CheckboxIndicator asChild>
-//         <Check weight="bold" />
-//       </CheckboxIndicator>
-//     </CheckboxContainer>
-//   )
-// }
+export function Checkbox(props: CheckboxProps) {
+  return (
+    <CheckboxContainer {...props}>
+      <CheckboxIndicator asChild>
+        <Check weight="bold" />
+      </CheckboxIndicator>
+    </CheckboxContainer>
+  )
+}