+++ packages/react/src/components/Heading.tsx
@@ -1,25 +1,42 @@
-import { styled } from '../styles'
+import styled, { css } from "styled-components"
+import { theme } from "../styles/themes"
 
-export const Heading = styled('h2', {
-  fontFamily: '$default',
-  lineHeight: '$shorter',
-  margin: 0,
-  color: '$gray100',
+const sizeVariants = {
+  sm: css`
+    font-size: ${theme.fontSizes.xs};
+  `,
+  md: css`
+    font-size: ${theme.fontSizes["2xl"]};
+  `,
+  lg: css`
+    font-size: ${theme.fontSizes["4xl"]};
+  `,
+  '2xl': css`
+    font-size: ${theme.fontSizes["5xl"]};
+  `,
+  '3xl': css`
+    font-size: ${theme.fontSizes["6xl"]};
+  `,
+  '4xl': css`
+    font-size: ${theme.fontSizes["7xl"]};
+  `,
+  '5xl': css`
+    font-size: ${theme.fontSizes["8xl"]};
+  `,
+  '6xl': css`
+    font-size: ${theme.fontSizes["9xl"]};
+  `,
+}
 
-  variants: {
-    size: {
-      sm: { fontSize: '$xl' },
-      md: { fontSize: '$2xl' },
-      lg: { fontSize: '$4xl' },
-      '2xl': { fontSize: '$5xl' },
-      '3xl': { fontSize: '$6xl' },
-      '4xl': { fontSize: '$7xl' },
-      '5xl': { fontSize: '$8xl' },
-      '6xl': { fontSize: '$9xl' },
-    },
-  },
+export const Heading = styled.h2<{ size: keyof typeof sizeVariants }>`
+  font-family: ${theme.fonts.default};
+  line-height: ${theme.lineHeights.shorter};
+  margin: 0;
+  color: ${theme.colors.gray100};
 
-  defaultVariants: {
-    size: 'md',
-  },
-})
+  ${({ size }) => sizeVariants[size]};
+`
+
+Heading.defaultProps = {
+  size: 'md',
+}