+++ packages/react/src/components/Text.tsx
@@ -1,30 +1,57 @@
-import { styled } from '../styles'
+import styled, { css } from "styled-components"
+import { theme } from "../styles/themes"
 
-export const Text = styled('p', {
-  fontFamily: '$default',
-  lineHeight: '$base',
-  margin: 0,
-  color: '$gray100',
+const sizeVariants = {
+  xxs: css`
+    font-size: ${theme.fontSizes.xxs};
+  `,
+  xs: css`
+    font-size: ${theme.fontSizes.xs};
+  `,
+  sm: css`
+    font-size: ${theme.fontSizes.sm};
+  `,
+  md: css`
+    font-size: ${theme.fontSizes.md};
+  `,
+  lg: css`
+    font-size: ${theme.fontSizes.lg};
+  `,
+  xl: css`
+    font-size: ${theme.fontSizes.xl};
+  `,
+  '2xl': css`
+    font-size: ${theme.fontSizes['2xl']};
+  `,
+  '4xl': css`
+    font-size: ${theme.fontSizes['4xl']};
+  `,
+  '5xl': css`
+    font-size: ${theme.fontSizes['5xl']};
+  `,
+  '6xl': css`
+    font-size: ${theme.fontSizes['6xl']};
+  `,
+  '7xl': css`
+    font-size: ${theme.fontSizes['7xl']};
+  `,
+  '8xl': css`
+    font-size: ${theme.fontSizes['8xl']};
+  `,
+  '9xl': css`
+    font-size: ${theme.fontSizes['9xl']};
+  `,
+}
 
-  variants: {
-    size: {
-      xxs: { fontSize: '$xxs' },
-      xs: { fontSize: '$xs' },
-      sm: { fontSize: '$sm' },
-      md: { fontSize: '$md' },
-      lg: { fontSize: '$lg' },
-      xl: { fontSize: '$xl' },
-      '2xl': { fontSize: '$2xl' },
-      '4xl': { fontSize: '$4xl' },
-      '5xl': { fontSize: '$5xl' },
-      '6xl': { fontSize: '$6xl' },
-      '7xl': { fontSize: '$7xl' },
-      '8xl': { fontSize: '$8xl' },
-      '9xl': { fontSize: '$9xl' },
-    },
-  },
+export const Text = styled.p<{ size?: keyof typeof sizeVariants}>`
+  font-family: ${theme.fonts.default};
+  line-height: ${theme.lineHeights.base};
+  margin: 0;
+  color: ${theme.colors.gray100};
 
-  defaultVariants: {
-    size: 'md',
-  },
-})
+  ${({ size = 'md' }) => sizeVariants[size]}
+`
+
+Text.defaultProps = {
+  size: 'md',
+}