+++ packages/react/src/styles/themes.ts
@@ -1,0 +1,21 @@
+import {
+  colors,
+  fonts,
+  fontSizes,
+  fontWeights,
+  lineHeights,
+  radii,
+  space
+} from '@gibbon-design/tokens'
+ 
+export const theme = {
+  colors,
+  fonts,
+  fontSizes,
+  fontWeights,
+  lineHeights,
+  radii,
+  space
+}
+
+export type ThemeType = typeof theme